feat: get the user data for profile
Install redux to store the user data. Get and write the user data in profile inputs and make them disabled, when the user clicks the edit button, inputs will become fillable and if the user clicks cancel or X button which is inside of input, all the inputs will reset to their default values which are stored in redux slice.