feat: finish eamil profile desktop version
add upload new photo function which displays new picture when the user chooses it from the computer, and when the user clicks save button the image gets posted to the server and after that client side does refetch to get the updated prof pic. Install react-toastify and add notifications to all request responses, either with errors or success messages and translate all of them. Also fix some bugs and refactor code a little bit.